Ticket #53: Incluindo recurso para dump do buffer de terminal.

+++ trunk/build.sh
@@ -39,21 +39,24 @@ fi
 SOURCE_PACKET=`rpm --eval="%{u2p:%{_srcrpmdir}}"`/$PACKAGE*.src.rpm
 PACKET_VERSION=`rpm -qp --qf '%{VERSION}-%{RELEASE}' $SOURCE_PACKET`
 
-echo Gerando copia do repositorio para a tag $PACKET_VERSION
-svn rm --force -m "Atualizando $PACKET_VERSION" http://$USER@suportelinux.df.bb.com.br/svn/$PACKAGE/tags/$PACKET_VERSION
-svn cp -m "Pacote $PACKET_VERSION" http://$USER@suportelinux.df.bb.com.br/svn/$PACKAGE/trunk http://$USER@suportelinux.df.bb.com.br/svn/$PACKAGE/tags/$PACKET_VERSION
+if [ "$1" != "--debug" ]; then
+   echo Gerando copia do repositorio para a tag $PACKET_VERSION
+   svn rm --force -m "Atualizando $PACKET_VERSION" http://$USER@suportelinux.df.bb.com.br/svn/$PACKAGE/tags/$PACKET_VERSION
+   svn cp -m "Pacote $PACKET_VERSION" http://$USER@suportelinux.df.bb.com.br/svn/$PACKAGE/trunk http://$USER@suportelinux.df.bb.com.br/svn/$PACKAGE/tags/$PACKET_VERSION
 
-echo "Enviando arquivo source para o servidor..."
-scp `rpm --eval="%{u2p:%{_srcrpmdir}}"`/$PACKAGE*.src.rpm $USER@suportelinux.df.bb.com.br:/dados/src/$VENDOR
-if [ "$?" != "0" ]; then
-   echo "Erro ao copiar o pacote fonte"
-   exit -1
-fi
+   echo "Enviando arquivo source para o servidor..."
+   scp `rpm --eval="%{u2p:%{_srcrpmdir}}"`/$PACKAGE*.src.rpm $USER@suportelinux.df.bb.com.br:/dados/src/$VENDOR
+   if [ "$?" != "0" ]; then
+      echo "Erro ao copiar o pacote fonte"
+      exit -1
+   fi
+   echo "Enviando arquivo binario o servidor..."
+   scp $RPMDIR/$RPMARCH/$PACKAGE*.rpm $USER@os2team.df.intrabb.bb.com.br:/home/matriz/pacotes/$VENDOR/bb
+   if [ "$?" != "0" ]; then
+      echo "Erro ao copiar o pacote binario"
+   fi
+   
+fi   
+
 mv -f `rpm --eval="%{u2p:%{_srcrpmdir}}"`/$PACKAGE*.src.rpm /tmp
-
-echo "Enviando arquivo binario o servidor..."
-scp $RPMDIR/$RPMARCH/$PACKAGE*.rpm $USER@os2team.df.intrabb.bb.com.br:/home/matriz/pacotes/$VENDOR/bb
-if [ "$?" = "0" ]; then
-   mv -f $RPMDIR/$RPMARCH/$PACKAGE*.rpm /tmp
-   echo "Binarios movidos para /tmp"
-fi
+mv -f $RPMDIR/$RPMARCH/$PACKAGE*.rpm /tmp
